+++ sdk/CHANGELOG.md
@@ -46,6 +46,7 @@
   Redesign `Client::relays` API (https://github.com/rust-nostr/nostr/pull/1231)
 - Redesign `Client::subscribe` and `Relay::subscribe` APIs (https://github.com/rust-nostr/nostr/pull/1232)
 - Redesign `Client::stream_events` and `Relay::stream_events` APIs (https://github.com/rust-nostr/nostr/pull/1233)
+- Redesign `Client::fetch_events` and `Relay::fetch_events` APIs (https://github.com/rust-nostr/nostr/pull/1234)
 
 ### Added
