+++ sdk/examples/aggregated-query.rs
@@ -2,6 +2,7 @@
 // Copyright (c) 2023-2025 Rust Nostr Developers
 // Distributed under the MIT software license
 
+use std::collections::HashMap;
 use std::time::Duration;
 
 use nostr_lmdb::NostrLmdb;
@@ -26,7 +27,10 @@ async fn main() -> Result<()> {
         .kind(Kind::TextNote)
         .limit(50);
     let stored_events = client.database().query(filter.clone()).await?;
-    let fetched_events = client.fetch_events(filter, Duration::from_secs(10)).await?;
+    let fetched_events = client
+        .fetch_events(filter)
+        .timeout(Duration::from_secs(10))
+        .await?;
     let events = stored_events.merge(fetched_events);
 
     for event in events.into_iter() {
@@ -41,14 +45,19 @@ async fn main() -> Result<()> {
 
     // Query events from relays
     let filter = Filter::new().author(public_key).kind(Kind::Metadata);
-    let fetched_events = client.fetch_events(filter, Duration::from_secs(10)).await?;
+    let fetched_events = client
+        .fetch_events(filter)
+        .timeout(Duration::from_secs(10))
+        .await?;
 
     // Add temp relay and fetch other events
     client.add_relay("wss://nostr.oxtr.dev").await?;
     client.connect_relay("wss://nostr.oxtr.dev").await?;
     let filter = Filter::new().kind(Kind::ContactList).limit(100);
+    let target = HashMap::from([("wss://nostr.oxtr.dev", filter)]);
     let fetched_events_from = client
-        .fetch_events_from(["wss://nostr.oxtr.dev"], filter, Duration::from_secs(10))
+        .fetch_events(target)
+        .timeout(Duration::from_secs(10))
         .await?;
     client.remove_relay("wss://nostr.oxtr.dev").force().await?;
