+++ sdk/examples/blacklist.rs
@@ -51,7 +51,10 @@ async fn main() -> Result<()> {
     let filter = Filter::new()
         .authors([muted_public_key, public_key])
         .kind(Kind::Metadata);
-    let events = client.fetch_events(filter, Duration::from_secs(10)).await?;
+    let events = client
+        .fetch_events(filter)
+        .timeout(Duration::from_secs(10))
+        .await?;
     println!("Received {} events.", events.len());
 
     assert_eq!(events.len(), 1);
