+++ sdk/examples/comment.rs
@@ -21,7 +21,8 @@ async fn main() -> Result<()> {
     let event_id =
         EventId::from_bech32("note1hrrgx2309my3wgeecx2tt6fl2nl8hcwl0myr3xvkcqpnq24pxg2q06armr")?;
     let events = client
-        .fetch_events(Filter::new().id(event_id), Duration::from_secs(10))
+        .fetch_events(Filter::new().id(event_id))
+        .timeout(Duration::from_secs(10))
         .await?;
 
     let comment_to = events.first().unwrap();
