+++ sdk/src/relay/api/mod.rs
@@ -1,7 +1,9 @@
+mod fetch_events;
 mod stream_events;
 mod subscribe;
 mod try_connect;
 
+pub use self::fetch_events::*;
 pub use self::stream_events::*;
 pub use self::subscribe::*;
 pub use self::try_connect::*;
