+++ sdk/src/relay/mod.rs
@@ -9,7 +9,6 @@ use std::time::Duration;
 
 use async_utility::time;
 use async_wsocket::ConnectionMode;
-use futures::StreamExt;
 use nostr_database::prelude::*;
 use tokio::sync::broadcast;
 
@@ -474,53 +473,12 @@ impl Relay {
     }
 
     /// Fetch events
-    pub async fn fetch_events<F>(
-        &self,
-        filters: F,
-        timeout: Option<Duration>,
-        policy: ReqExitPolicy,
-    ) -> Result<Events, Error>
+    #[inline]
+    pub fn fetch_events<F>(&self, filters: F) -> FetchEvents
     where
         F: Into<Vec<Filter>>,
     {
-        let filters: Vec<Filter> = filters.into();
-
-        // Construct a new events collection
-        let mut events: Events = if filters.len() == 1 {
-            // SAFETY: this can't panic because the filters are already verified that list isn't empty.
-            let filter: &Filter = &filters[0];
-            Events::new(filter)
-        } else {
-            // More than a filter, so we can't ensure to respect the limit -> construct a default collection.
-            Events::default()
-        };
-
-        // Stream events
-        let mut stream = self
-            .stream_events(filters)
-            .maybe_timeout(timeout)
-            .policy(policy)
-            .await?;
-
-        while let Some(res) = stream.next().await {
-            // Get event from the result
-            let event: Event = res?;
-
-            // Use force insert here!
-            // Due to the configurable REQ exit policy, the user may want to wait for events after EOSE.
-            // If the filter has a limit, the force insert allows adding events post-EOSE.
-            //
-            // For example, if we use `Events::insert` here,
-            // if the filter is '{"kinds":[1],"limit":3}' and the policy `ReqExitPolicy::WaitForEventsAfterEOSE(1)`,
-            // the events collection will discard 1 event because the filter limit is 3 and the total received events are 4.
-            //
-            // Events::force_insert automatically increases the capacity if needed, without discarding events.
-            //
-            // LOOKUP_ID: EVENTS_FORCE_INSERT
-            events.force_insert(event);
-        }
-
-        Ok(events)
+        FetchEvents::new(self, filters.into())
     }
 
     /// Count events
@@ -656,31 +614,6 @@ mod tests {
         opts: RelayOptions,
     ) -> Relay {
         Relay::builder(url).database(database).opts(opts).build()
-    }
-
-    /// Setup public (without NIP42 auth) relay with N events to test event fetching
-    ///
-    /// **Adds ONLY text notes**
-    async fn setup_event_fetching_relay(num_events: usize) -> (Relay, MockRelay) {
-        // Mock relay
-        let mock = MockRelay::run().await.unwrap();
-        let url = mock.url().await;
-
-        let relay = new_relay(url, RelayOptions::default());
-        relay.connect();
-
-        // Signer
-        let keys = Keys::generate();
-
-        // Send some events
-        for i in 0..num_events {
-            let event = EventBuilder::text_note(i.to_string())
-                .sign_with_keys(&keys)
-                .unwrap();
-            relay.send_event(&event).await.unwrap();
-        }
-
-        (relay, mock)
     }
 
     async fn setup_subscription_relay() -> (SubscriptionId, Relay, MockRelay) {
@@ -1071,302 +1004,6 @@ mod tests {
     }
 
     #[tokio::test]
-    async fn test_fetch_events_ban_relay() {
-        // Mock relay
-        let opts = LocalRelayTestOptions {
-            unresponsive_connection: None,
-            send_random_events: true,
-        };
-        let mock = MockRelay::run_with_opts(opts).await.unwrap();
-        let url = mock.url().await;
-
-        let relay: Relay = new_relay(
-            url,
-            RelayOptions::default()
-                .verify_subscriptions(true)
-                .ban_relay_on_mismatch(true),
-        );
-
-        assert_eq!(relay.status(), RelayStatus::Initialized);
-
-        relay
-            .try_connect()
-            .timeout(Duration::from_secs(3))
-            .await
-            .unwrap();
-
-        assert_eq!(relay.status(), RelayStatus::Connected);
-
-        let filter = Filter::new().kind(Kind::Metadata);
-        relay
-            .fetch_events(
-                filter,
-                Some(Duration::from_secs(3)),
-                ReqExitPolicy::ExitOnEOSE,
-            )
-            .await
-            .unwrap();
-
-        assert_eq!(relay.status(), RelayStatus::Banned);
-
-        assert!(!relay.inner.is_running());
-    }
-
-    #[tokio::test]
-    async fn test_nip42_send_event() {
-        // Mock relay
-        let opts = LocalRelayBuilderNip42 {
-            mode: LocalRelayBuilderNip42Mode::Write,
-        };
-        let mock = LocalRelay::builder().nip42(opts).build();
-        mock.run().await.unwrap();
-        let url = mock.url().await;
-
-        let relay: Relay = new_relay(url, RelayOptions::default());
-
-        relay.inner.state.automatic_authentication(true);
-
-        relay.connect();
-
-        // Signer
-        let keys = Keys::generate();
-
-        // Send as unauthenticated (MUST return error)
-        let event = EventBuilder::text_note("Test")
-            .sign_with_keys(&keys)
-            .unwrap();
-        let err = relay.send_event(&event).await.unwrap_err();
-        if let Error::RelayMessage(msg) = err {
-            assert_eq!(
-                MachineReadablePrefix::parse(&msg).unwrap(),
-                MachineReadablePrefix::AuthRequired
-            );
-        } else {
-            panic!("Unexpected error");
-        }
-
-        // Set a signer
-        relay.inner.state.set_signer(keys.clone()).await;
-
-        // Send as authenticated
-        let event = EventBuilder::text_note("Test")
-            .sign_with_keys(&keys)
-            .unwrap();
-        assert!(relay.send_event(&event).await.is_ok());
-    }
-
-    #[tokio::test]
-    async fn test_nip42_fetch_events() {
-        // Mock relay
-        let opts = LocalRelayBuilderNip42 {
-            mode: LocalRelayBuilderNip42Mode::Read,
-        };
-        let mock = LocalRelay::builder().nip42(opts).build();
-        mock.run().await.unwrap();
-        let url = mock.url().await;
-
-        let relay: Relay = new_relay(url, RelayOptions::default());
-
-        relay.connect();
-
-        // Signer
-        let keys = Keys::generate();
-
-        // Send an event
-        let event = EventBuilder::text_note("Test")
-            .sign_with_keys(&keys)
-            .unwrap();
-        relay.send_event(&event).await.unwrap();
-
-        let filter = Filter::new().kind(Kind::TextNote).limit(3);
-
-        // Disable NIP42 auto auth
-        relay.inner.state.automatic_authentication(false);
-
-        // Unauthenticated fetch (MUST return error)
-        let err = relay
-            .fetch_events(
-                filter.clone(),
-                Some(Duration::from_secs(5)),
-                ReqExitPolicy::ExitOnEOSE,
-            )
-            .await
-            .unwrap_err();
-        match err {
-            Error::RelayMessage(msg) => {
-                assert_eq!(
-                    MachineReadablePrefix::parse(&msg).unwrap(),
-                    MachineReadablePrefix::AuthRequired
-                );
-            }
-            e => panic!("Unexpected error: {e}"),
-        }
-
-        // Enable NIP42 auto auth
-        relay.inner.state.automatic_authentication(true);
-
-        // Unauthenticated fetch (MUST return error)
-        let err = relay
-            .fetch_events(
-                filter.clone(),
-                Some(Duration::from_secs(5)),
-                ReqExitPolicy::ExitOnEOSE,
-            )
-            .await
-            .unwrap_err();
-        assert!(matches!(err, Error::AuthenticationFailed));
-
-        // Set a signer
-        relay.inner.state.set_signer(keys).await;
-
-        // Authenticated fetch
-        let res = relay
-            .fetch_events(
-                filter,
-                Some(Duration::from_secs(5)),
-                ReqExitPolicy::ExitOnEOSE,
-            )
-            .await;
-        assert!(res.is_ok());
-    }
-
-    #[tokio::test]
-    async fn test_fetch_events_exit_on_eose() {
-        let (relay, _mock) = setup_event_fetching_relay(5).await;
-
-        // Exit on EOSE
-        let events = relay
-            .fetch_events(
-                Filter::new().kind(Kind::TextNote),
-                Some(Duration::from_secs(5)),
-                ReqExitPolicy::ExitOnEOSE,
-            )
-            .await
-            .unwrap();
-        assert_eq!(events.len(), 5);
-
-        // Exit on EOSE
-        let events = relay
-            .fetch_events(
-                Filter::new().kind(Kind::TextNote).limit(3),
-                Some(Duration::from_secs(5)),
-                ReqExitPolicy::ExitOnEOSE,
-            )
-            .await
-            .unwrap();
-        assert_eq!(events.len(), 3);
-    }
-
-    #[tokio::test]
-    async fn test_fetch_events_wait_for_events() {
-        let (relay, _mock) = setup_event_fetching_relay(5).await;
-
-        let events = relay
-            .fetch_events(
-                Filter::new().kind(Kind::TextNote),
-                Some(Duration::from_secs(15)),
-                ReqExitPolicy::WaitForEvents(2),
-            )
-            .await
-            .unwrap();
-        assert_eq!(events.len(), 2); // Requested all text notes but exit after receive 2
-
-        // Task to send additional event
-        let r = relay.clone();
-        tokio::spawn(async move {
-            tokio::time::sleep(Duration::from_secs(2)).await;
-
-            // Signer
-            let keys = Keys::generate();
-
-            // Build and send event
-            let event = EventBuilder::metadata(&Metadata::new().name("Test"))
-                .sign_with_keys(&keys)
-                .unwrap();
-            r.send_event(&event).await.unwrap();
-        });
-
-        let events = relay
-            .fetch_events(
-                Filter::new().kind(Kind::Metadata),
-                Some(Duration::from_secs(5)),
-                ReqExitPolicy::WaitForEvents(1),
-            )
-            .await
-            .unwrap();
-        assert_eq!(events.len(), 1);
-    }
-
-    #[tokio::test]
-    async fn test_fetch_events_wait_for_events_after_eose() {
-        let (relay, _mock) = setup_event_fetching_relay(10).await;
-
-        // Task to send additional events
-        let r = relay.clone();
-        tokio::spawn(async move {
-            // Signer
-            let keys = Keys::generate();
-
-            // Send more events
-            for _ in 0..2 {
-                // Sleep
-                tokio::time::sleep(Duration::from_secs(2)).await;
-
-                // Build and send event
-                let event = EventBuilder::text_note("Additional")
-                    .sign_with_keys(&keys)
-                    .unwrap();
-                r.send_event(&event).await.unwrap();
-            }
-        });
-
-        let events = relay
-            .fetch_events(
-                Filter::new().kind(Kind::TextNote).limit(3),
-                Some(Duration::from_secs(15)),
-                ReqExitPolicy::WaitForEventsAfterEOSE(2),
-            )
-            .await
-            .unwrap();
-        assert_eq!(events.len(), 5); // 3 events received until EOSE + 2 new events
-    }
-
-    #[tokio::test]
-    async fn test_fetch_events_wait_for_duration_after_eose() {
-        let (relay, _mock) = setup_event_fetching_relay(5).await;
-
-        // Task to send additional events
-        let r = relay.clone();
-        tokio::spawn(async move {
-            tokio::time::sleep(Duration::from_secs(2)).await;
-
-            // Signer
-            let keys = Keys::generate();
-
-            // Send more events
-            for _ in 0..2 {
-                // Build and send event
-                let event = EventBuilder::text_note("Additional")
-                    .sign_with_keys(&keys)
-                    .unwrap();
-                r.send_event(&event).await.unwrap();
-
-                tokio::time::sleep(Duration::from_secs(2)).await;
-            }
-        });
-
-        let events = relay
-            .fetch_events(
-                Filter::new().kind(Kind::TextNote),
-                Some(Duration::from_secs(15)),
-                ReqExitPolicy::WaitDurationAfterEOSE(Duration::from_secs(3)),
-            )
-            .await
-            .unwrap();
-        assert_eq!(events.len(), 6); // 5 events received until EOSE + 1 new events
-    }
-
-    #[tokio::test]
     async fn test_unsubscribe() {
         let (id, relay, _mock) = setup_subscription_relay().await;
 
@@ -1538,11 +1175,8 @@ mod tests {
         // Test wake up when fetch events
         let filter = Filter::new().kind(Kind::TextNote);
         let _ = relay
-            .fetch_events(
-                filter,
-                Some(Duration::from_secs(10)),
-                ReqExitPolicy::ExitOnEOSE,
-            )
+            .fetch_events(filter)
+            .timeout(Duration::from_secs(10))
             .await
             .unwrap();
         assert_eq!(relay.status(), RelayStatus::Connected);
